pyproject.toml
[project]
name = "eteragproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "fastapi>=0.128.0",
    "inngest>=0.5.13",
    "langchain-openai>=1.1.6",
    "llama-index-core>=0.14.12",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.14.0",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "ragas>=0.4.2",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
